extract wishlist icon in another component
