+++ src/components/Card/Card.css
@@ -2,22 +2,6 @@
   width: 100%;
 }
 
-.card__fav {
-  position: absolute;
-  top: 0.5rem;
-  right: 0.5rem;
-
-  width: 1.75rem;
-  height: 1.75rem;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  color: var(--red);
-  background: var(--white);
-  border-radius: 50%;
-  cursor: pointer;
-}
-
 .card__subtitle--canceled {
   font-weight: 400;
   font-size: 1rem;
@@ -43,10 +27,6 @@
   font-size: 1.5rem;
   line-height: 2.25rem;
   text-decoration: line-through;
-}
-
-.card--horizontal .card__fav {
-  display: none;
 }
 
 .card--horizontal .quantity__info,
