+++ src/components/Card/Card.js
@@ -1,35 +1,15 @@
 import React from "react";
-import { useNavigate } from "react-router-dom";
-import {
-  useAddToWishlistMutation,
-  useRemoveFromWishlistMutation,
-  useUserQuery
-} from "../../hooks";
+import { useLocation } from "react-router-dom";
+import { WishlistIconButton } from "../../components";
 import "./Card.css";
 
 const Card = ({ product }) => {
-  const { mutate: addToWishlist } = useAddToWishlistMutation();
-  const { mutate: removeFromWishlist } = useRemoveFromWishlistMutation();
-  const { data } = useUserQuery();
-  const { user } = data ?? {};
-  const navigate = useNavigate();
-
+  const { pathname } = useLocation();
   const { id, name, price, discount, pictureLink, cartItem } = product;
 
-  const wishlisted = user?.wishlist?.some(item => item.id === id);
-
-  const wishlistHandler = () => {
-    if (!user?.id) {
-      return navigate("/signup");
-    }
-    return wishlisted ? removeFromWishlist(id) : addToWishlist(id);
-  };
-
   return (
-    <div className="card">
-      <div className="card__fav" onClick={wishlistHandler}>
-        <i className={`${wishlisted ? "fa-solid" : "fa-regular"} fa-heart`}></i>
-      </div>
+    <div className={`card ${pathname === "/cart" ? "card--horizontal" : ""}`}>
+      <WishlistIconButton productId={id} />
       <img className="card__image" src={pictureLink} alt={name} />
       <div className="card__content">
         <h2 className="card__title">{name}</h2>
