+++ src/components/WishlistIconButton/WishlistIconButton.css
@@ -1,0 +1,19 @@
+.card__fav {
+  position: absolute;
+  top: 0.5rem;
+  right: 0.5rem;
+
+  width: 1.75rem;
+  height: 1.75rem;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  color: var(--red);
+  background: var(--white);
+  border-radius: 50%;
+  cursor: pointer;
+}
+
+.card--horizontal .card__fav {
+  display: none;
+}
