+++ src/components/index.js
@@ -5,3 +5,4 @@ export { Header } from "./Header/Header";
 export { InputField } from "./InputField/InputField";
 export { NewArrivals } from "./NewArrivals/NewArrivals";
 export { ProductList } from "./ProductList/ProductList";
+export { WishlistIconButton } from "./WishlistIconButton/WishlistIconButton";
